Extract school form schemas and add class dialog
Move inline Zod schemas and defaults into schema.ts for onboarding
create/join. Add CreateSchoolClassDialog component and refactor the
classes header to use it. Rename dialog prop setOpen → setOpenAction
across materials and people components and update call sites. Add
brief JSDoc comments for form and dialog modules.
